backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
# mysqlclient optional - using PyMySQL driver instead
PyMySQL==1.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.0.1
anthropic==0.18.0
psutil==5.9.6
cryptography==41.0.7
bcrypt==4.1.2
# Agent System
APScheduler==3.11.0
schedule==1.2.2
requests==2.32.3
python-telegram-bot==20.7
